requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
creategitignore==0.2.1
Django==4.2.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
idna==3.4
inflection==0.5.1
joblib==1.2.0
jsonschema==4.17.3
newsapi-python==0.2.7
nltk==3.8.1
numpy==1.24.3
praw==7.7.0
prawcore==2.3.0
pyrsistent==0.19.3
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
soupsieve==2.4.1
sqlparse==0.4.4
threadpoolctl==3.1.0
tqdm==4.65.0
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.0.2
utils==1.0.1
websocket-client==1.5.2
whitenoise==6.4.0
